% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_summary.R
\name{cov_summary}
\alias{cov_summary}
\title{Coverage summary}
\usage{
cov_summary(
  cov.dir = "./data/",
  output.dir = "",
  cov.suffix = "cov.gz",
  chr.to.keep = c(1:22, "X", "Y", "y", "x"),
  anno = "NULL",
  run.parallel = FALSE,
  n.cores = NA
)
}
\arguments{
\item{cov.dir}{directory containing .cov files.}

\item{output.dir}{Directory to save results table. Default is location of 'cov.dir' provided.}

\item{cov.suffix}{Ending pattern to match and identify files. Default is 'cov.gz'.}

\item{chr.to.keep}{Select chromosomes to keep. Default = c(1:22,"X", "Y", "y", "x")}

\item{anno}{Annotation file. Needs to be file that has a dataframe with columns 'chr', 'start' and 'end' at the very least. The script will then summary statistics for all the whole cov (filtered by selected chr's) + cytosines within the ranges provided in this file. Default is 'NULL'.}

\item{run.parallel}{Logical. Whether or not to run using parallel processing. Should work for both Unix and windows... (untested in mac)}

\item{n.cores}{Numeric. Optional to specify with parallel processing. If n.cores = NA & run.parallel = TRUE, the script will identify total available cores and use 1-total. Advise not using 'all cores' if it's on your local machine.}
}
\value{
genome fasta
}
\description{
Summarise coverage (.cov) files generated by bismark in a given directory. An annotation file, containing specific genomic co-ordinates per row, can be given to provide summary level of cytosine within the ranges provided.
}
\examples{
cov_summary(cov.dir = system.file("extdata", package = "covSummary"), output.dir = "./test")
}
