% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SINE.Alu_anno.R
\name{Create.Filtered.Annotation}
\alias{Create.Filtered.Annotation}
\title{Filter fasta for SINE-Alu}
\usage{
Create.Filtered.Annotation(
  anno.file = "./Annotations/hg38.fa.out.gz",
  output.dir = "./Annotations/",
  filtered.anno.name = "SINE.Alu.anno.txt"
)
}
\arguments{
\item{anno.file}{Annotation file}

\item{output.dir}{Directory to save output}

\item{filtered.anno.name}{Output filename.}
}
\value{
Filtered annotation for SINE-Alu sites.
}
\description{
Filter genome fasta for SINE-Alu annotations
}
\examples{
Download.Annotation(annotation.dir = "./anno/");
}
