% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SINE.Alu_anno.R
\name{Download.Annotation}
\alias{Download.Annotation}
\title{Download hg38 fasta}
\usage{
Download.Annotation(
  URL = "https://hgdownload.cse.ucsc.edu/goldenPath/hg38/bigZips/latest/hg38.fa.out.gz",
  annotation.dir = "./Annotations/"
)
}
\arguments{
\item{URL}{ftp path of genome fasta}

\item{annotation.dir}{Directoy to save downloaded fasta}
}
\value{
genome fasta
}
\description{
Download latest hg38 genome fasta from UCSC
}
\examples{
Download.Annotation(annotation.dir = "./anno/");
}
